+++ XGame/XSrc/XEngine/XPlatform/math/XMathUtil.h
@@ -42,8 +42,8 @@ void XMathVector3TransformCoord(const XMatrix& m, const XVector3& v, XVector3& o
 
 int XRandomInt();
 int XRandomInt(int aMin, int bMax);
-int XRandomFloat();
-int XRandomFloat(float aMin, float bMax);
+float XRandomFloat();
+float XRandomFloat(float aMin, float bMax);
 void XRandomVector3();
 
 #endif // XMathUtil
